Commit No 10 - Enemy
The enemy can cause damage through various methods, one of them is shown
in the image uploaded with this commit, the enemy will shoot lasers at
the player and it will decrease the health of the player. Other methods
are: decreasing the score when picking up wrong objects, and timer going
down.